Container and Resolver interfaces
